Move key creation function to separate module
